Proof of concept Transformable sealed trait hierarchy and single TransformableParser
