+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/Server.scala
@@ -11,7 +11,6 @@ import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.transformer.modules.{
-  TransformableParserModule,
   TransformerWorkerModule
 }
 
@@ -28,7 +27,6 @@ class Server extends HttpServer {
     SQSConfigModule,
     SNSConfigModule,
     SNSClientModule,
-    TransformableParserModule,
     TransformerWorkerModule
   )
   override def configureHttp(router: HttpRouter) {
