--- catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/parsers/MiroParser.scala
@@ -1,13 +1,0 @@
-package uk.ac.wellcome.transformer.parsers
-
-import com.twitter.inject.Logging
-import uk.ac.wellcome.models.transformable.miro.MiroTransformable
-import uk.ac.wellcome.utils.JsonUtil
-
-import scala.util.Try
-
-class MiroParser extends TransformableParser[MiroTransformable] with Logging {
-  override def readFromRecord(
-    transformableAsJson: String): Try[MiroTransformable] =
-    JsonUtil.fromJson[MiroTransformable](transformableAsJson)
-}
