+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/transformer/parsers/SierraParserTest.scala
@@ -1,8 +1,7 @@
 package uk.ac.wellcome.transformer.parsers
 
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.models.MergedSierraRecord
-import uk.ac.wellcome.models.transformable.Transformable
+import uk.ac.wellcome.models.transformable.{MergedSierraRecord, Transformable}
 import uk.ac.wellcome.transformer.utils.TransformableSQSMessageUtils
 import uk.ac.wellcome.utils.JsonUtil
 
@@ -16,7 +15,7 @@ class SierraParserTest
     val id = "000"
     val sqsMessage = createValidEmptySierraBibSQSMessage(id)
 
-    val sierraParser = new SierraParser
+    val sierraParser = new TransformableParser
 
     val triedSierraTransformable =
       sierraParser.extractTransformable(sqsMessage)
