+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/transformer/utils/TransformableSQSMessageUtils.scala
@@ -3,13 +3,8 @@ package uk.ac.wellcome.transformer.utils
 import java.time.Instant
 
 import uk.ac.wellcome.models.aws.SQSMessage
-import uk.ac.wellcome.models.transformable.miro.MiroTransformable
-import uk.ac.wellcome.models.{
-  CalmTransformable,
-  MergedSierraRecord,
-  SierraBibRecord,
-  SierraItemRecord
-}
+import uk.ac.wellcome.models.transformable.{CalmTransformable, MergedSierraRecord, MiroTransformable}
+import uk.ac.wellcome.models.{SierraBibRecord, SierraItemRecord}
 import uk.ac.wellcome.utils.JsonUtil
 
 trait TransformableSQSMessageUtils {
