+++ common/src/main/scala/uk/ac/wellcome/models/CalmTransformable.scala
@@ -7,7 +7,7 @@ import scala.util.Try
 
 case class CalmTransformableData(
   AccessStatus: Array[String]
-) extends Transformable {
+) {
   def transform: Try[Option[Work]] = Try {
     // TODO: Fill in proper data here
     Some(
@@ -26,26 +26,3 @@ case class CalmTransformableData(
   }
 }
 
-//TODO add some tests around transformation
-case class CalmTransformable(
-  RecordID: String,
-  RecordType: String,
-  AltRefNo: String,
-  RefNo: String,
-  data: String,
-  ReindexShard: String = "default",
-  ReindexVersion: Int = 0
-) extends Transformable
-    with Reindexable[String] {
-
-  val id: ItemIdentifier[String] = ItemIdentifier(
-    HashKey("RecordID", RecordID),
-    RangeKey("RecordType", RecordType)
-  )
-
-  def transform: Try[Option[Work]] =
-    JsonUtil
-      .fromJson[CalmTransformableData](data)
-      .flatMap(_.transform)
-
-}
