+++ common/src/test/scala/uk/ac/wellcome/test/utils/MiroTransformableWrapper.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.test.utils
 
 import org.scalatest.{Matchers, Suite}
 import uk.ac.wellcome.models.Work
-import uk.ac.wellcome.models.transformable.miro.MiroTransformable
+import uk.ac.wellcome.models.transformable.MiroTransformable
 
 /** MiroTransformable looks for several fields in the source JSON -- if they're
   *  missing or have the wrong values, it rejects the record.
