+++ sierra_adapter/sierra_item_merger/src/test/scala/uk/ac/wellcome/platform/sierra_item_merger/services/SierraItemMergerUpdaterServiceTest.scala
@@ -1,18 +1,18 @@
 package uk.ac.wellcome.platform.sierra_item_merger.services
 
 import com.gu.scanamo.Scanamo
+import com.gu.scanamo.syntax._
 import org.mockito.Matchers.any
 import org.mockito.Mockito.when
 import org.scalatest.FunSpec
 import uk.ac.wellcome.metrics.MetricsSender
-import uk.ac.wellcome.models.MergedSierraRecord
 import uk.ac.wellcome.models.aws.DynamoConfig
+import uk.ac.wellcome.models.transformable.MergedSierraRecord
+import uk.ac.wellcome.platform.sierra_adapter.dynamo.MergedSierraRecordDao
 import uk.ac.wellcome.platform.sierra_item_merger.utils.SierraItemMergerTestUtil
 import uk.ac.wellcome.dynamo._
-import uk.ac.wellcome.platform.sierra_adapter.dynamo.MergedSierraRecordDao
-import scala.concurrent.Future
 
-import com.gu.scanamo.syntax._
+import scala.concurrent.Future
 
 class SierraItemMergerUpdaterServiceTest
     extends FunSpec
